reorganized home page components
