+++ src/app/page.tsx
@@ -1,6 +1,6 @@
-import Hero from "@/components/Hero";
-import Experiences from "@/components/Experiences";
-import Projects from "@/components/Projects";
+import Hero from "@/components/home/Hero/Hero";
+import Experiences from "@/components/home/Experiences/Experiences";
+import Projects from "@/components/home/Projects/Projects";
 import heroData from "@/content/hero.json";
 import experienceData from "@/content/experience.json";
 import codeData from "@/content/code.json";
